Moved class unrelated to Gradle out of 'gradle' package
