+++ src/main/groovy/org/shipkit/internal/util/TemplateResolver.java
@@ -1,6 +1,4 @@
-package org.shipkit.internal.gradle;
-
-import org.shipkit.internal.util.ArgumentValidation;
+package org.shipkit.internal.util;
 
 import java.util.HashMap;
 import java.util.Map;
