+++ src/test/groovy/org/shipkit/internal/util/TemplateResolverTest.groovy
@@ -1,4 +1,4 @@
-package org.shipkit.internal.gradle
+package org.shipkit.internal.util
 
 import spock.lang.Specification
